wrapper over duckdb
