+++ packages/railtracks/tests/unit_tests/query/test_connect.py
@@ -1,4 +1,4 @@
-"""Tests for railtracks.query."""
+"""Tests for railtracks.query.connect — DuckDB views + EventQuery."""
 
 from __future__ import annotations
 
@@ -10,59 +10,9 @@ import pytest
 
 pytest.importorskip("duckdb")
 
-from railtracks.query import EventQuery, connect, list_namespaces
+from railtracks.query import connect
 
-
-# Real-shaped events for the fixtures. Envelope matches events.jsonl:
-# event_id, event_type, scope_type, scope_id, parent_scope_id,
-# parent_event_id, stamp (struct or ISO string — both are valid JSON), payload.
-
-LLM_CALL = {
-    "event_id": "evt_llm_1",
-    "event_type": "llm.call",
-    "scope_type": "session",
-    "scope_id": "sess_a1",
-    "parent_scope_id": None,
-    "parent_event_id": None,
-    "stamp": {"ts": "2026-07-10T14:00:00.000Z", "seq": 1},
-    "payload": {
-        "session_id": "sess_a1",
-        "node_id": "node_root",
-        "model": "claude-opus-4-7",
-        "provider": "Anthropic",
-        "input_tokens": 812,
-        "output_tokens": 204,
-        "total_cost": 0.005,
-        "latency_s": 1.2,
-    },
-}
-
-NODE_START = {
-    "event_id": "evt_node_1",
-    "event_type": "node.start",
-    "scope_type": "session",
-    "scope_id": "sess_a1",
-    "parent_scope_id": None,
-    "parent_event_id": None,
-    "stamp": {"ts": "2026-07-10T14:00:00.100Z", "seq": 2},
-    "payload": {
-        "session_id": "sess_a1",
-        "node_id": "node_root",
-        "node_type": "Agent",
-        "node_name": "Root",
-    },
-}
-
-CUSTOM_NS_EVENT = {
-    "event_id": "evt_custom_1",
-    "event_type": "custom.thing",
-    "scope_type": "session",
-    "scope_id": "sess_a1",
-    "parent_scope_id": None,
-    "parent_event_id": None,
-    "stamp": {"ts": "2026-07-10T14:00:00.200Z", "seq": 3},
-    "payload": {"foo": "bar", "children": [{"x": 1}]},
-}
+from ._events import CUSTOM_NS_EVENT, LLM_CALL, NODE_START
 
 
 def _write(path: Path, events: list[dict]) -> None:
@@ -182,14 +132,6 @@ class TestEmptyNamespaces:
             assert q.namespaces == []
             assert q.namespaces_missing == []
             assert _tables(q.con) == {"events"}
-
-
-class TestListNamespaces:
-    def test_returns_sorted_distinct_union_of_samples_and_data(self, tmp_path: Path):
-        f = tmp_path / "e.jsonl"
-        _write(f, [LLM_CALL, CUSTOM_NS_EVENT])
-        # samples floor: session, node, llm. data adds: llm (dup), custom.
-        assert list_namespaces(f) == sorted({"session", "node", "llm", "custom"})
 
 
 class TestPayloadColumnBehavior:
